Implementación de persistencia a base de datos para devolución de estadísticas.

+++ dna-checker/src/main/java/com/exercise/mutant/webservices/dnachecker/model/MutantStats.java
@@ -9,39 +9,39 @@ public class MutantStats {
 	/**
 	 * Atributo que representa la cantidad de ADN mutante procesado
 	 */
-	private int count_mutant_dna;
+	private long count_mutant_dna;
 	
 	/**
 	 * Atributo que representa la cantidad de ADN humano procesado
 	 */
-	private int count_human_dna;
+	private long count_human_dna;
 	
 
 	/**
 	 * @return Retorna la cantidad de ADN mutante
 	 */
-	public int getCount_mutant_dna() {
+	public long getCount_mutant_dna() {
 		return count_mutant_dna;
 	}
 
 	/**
 	 * @param Establece la cantidad de ADN mutante
 	 */
-	public void setCount_mutant_dna(int count_mutant_dna) {
+	public void setCount_mutant_dna(long count_mutant_dna) {
 		this.count_mutant_dna = count_mutant_dna;
 	}
 
 	/**
 	 * @return Retorna la cantidad de ADN humano
 	 */
-	public int getCount_human_dna() {
+	public long getCount_human_dna() {
 		return count_human_dna;
 	}
 
 	/**
 	 * @param Establece la cantidad de ADN humano
 	 */
-	public void setCount_human_dna(int count_human_dna) {
+	public void setCount_human_dna(long count_human_dna) {
 		this.count_human_dna = count_human_dna;
 	}
